apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
tzdata==2025.1
